+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortalLogController.test.ts
@@ -43,6 +43,7 @@ describe('PortalLogController', () => {
             body: [{
                 level: 'error',
                 message: 'This is the error message',
+                path: '/test',
                 portalAppName: 'App2'
             }]
         };
@@ -61,7 +62,7 @@ describe('PortalLogController', () => {
         const controller = new PortalLogController(pluginRegistry);
         await controller.log(req, res);
 
-        expect(errorMessage).toBe('This is the error message [App2 v2.2.4]');
+        expect(errorMessage).toBe('This is the error message [Page: /test] [App: App2 v2.2.4]');
         expect(context).toEqual({
             portalAppName: 'App2',
             portalAppVersion: '2.2.4'